+++ examples/01_plotting/plot_3d_map_projection.py
@@ -2,15 +2,14 @@
 Making a surface plot of a 3D statistical map
 =============================================
 
-Download a statistical map from neurovault and a cortical mesh (fsaverage),
+Download a statistical map from Neurovault and a cortical mesh (fsaverage),
 project the 3D data on the mesh and display a surface plot.
-"""
 
-from matplotlib import pyplot as plt
+"""
 
 import nilearn.image
 import nilearn.datasets
-from nilearn.plotting import surf_plotting
+import nilearn.plotting
 
 ##############################################################################
 # Get a statistical map
@@ -29,13 +28,13 @@ fsaverage = nilearn.datasets.fetch_surf_fsaverage5()
 # Sample the 3D data around each node of the mesh
 # -----------------------------------------------
 
-texture = surf_plotting.niimg_to_surf_data(image, fsaverage.pial_left, kind='line')
+texture = nilearn.plotting.niimg_to_surf_data(image, fsaverage.pial_left)
 
 
 ##############################################################################
 # Plot the result
 # ---------------
 
-surf_plotting.plot_surf_stat_map(fsaverage.infl_left, texture, cmap='bwr')
+nilearn.plotting.plot_surf_stat_map(fsaverage.infl_left, texture, cmap='bwr')
 
-plt.show()
+nilearn.plotting.show()